+++ src/utils/cache.js
@@ -1,5 +1,6 @@
 const express = require('express')
 const path = require('path')
+const mCache = require('memory-cache')
 
 const ENV = process.env.NODE_ENV
 
@@ -24,6 +25,27 @@ const cacheOptions = {
     }
 }
 
-// TODO selected API caching
+exports.cache = function(duration) {
+    return function (req, res, next) {
+        const key = '__express__' + req.originalUrl || req.url
+        // check if the response is cached
+        const cachedResponse = mCache.get(key)
+        if (cachedResponse) {
+            // if so, send it
+            res.send(cachedResponse)
+            return
+        }
 
+        // patch in a new send function that puts the result
+        // of the request into the cache
+        res.sendResponse = res.send
+        res.send = (body) => {
+            mCache.put(key, body, duration * 1000)
+            res.sendResponse(body)
+        }
+        next()
+    }
+}
+
+// TODO selected API caching
 exports.staticFilesCache = express.static(staticPath, cacheOptions)
