+++ src/utils/helpers.js
@@ -1,27 +1,5 @@
 const { APIError, ERRORS } = require('../error')
 const { PERMISSIONS_MOD } = require('../models/constants')
-const mCache = require('memory-cache')
-
-exports.cache = function(duration) {
-    return function (req, res, next) {
-        const key = '__express__' + req.originalUrl || req.url
-        // check if the response is cached
-        const cachedResponse = mCache.get(key)
-        if (cachedResponse) {
-            // if so, send it
-            res.send(cachedResponse)
-            return
-        } else {
-            // patch in a new send function that puts the result of the request into the cache
-            res.sendResponse = res.send
-            res.send = (body) => {
-                mCache.put(key, body, duration * 1000)
-                res.sendResponse(body)
-            }
-        }
-        next()
-    }
-}
 
 /**
  * Takes a successful GET response and creates
